+++ packages/components-js/tests/vrt/specs/link-pure.vrt.ts
@@ -30,7 +30,10 @@ describe('Link Pure', () => {
       await vrt.test('link-pure-states', async () => {
         const page = await vrt.getPage();
 
-        const head = `<style type="text/css">p-link-pure:not(:last-child) { margin-right: 0.5rem; } div div:not(:first-of-type) { margin-top: 0.5rem }</style>`;
+        const head = `<style type="text/css">
+          p-link-pure:not(:last-child) { margin-right: 0.5rem; }
+          div div:not(:first-of-type) { margin-top: 0.5rem; }
+        </style>`;
 
         const getElementsMarkup: GetThemedMarkup = (theme) => `
           <div>
@@ -77,12 +80,13 @@ describe('Link Pure', () => {
 
         await setContentWithDesignSystem(page, getThemedBodyMarkup(getElementsMarkup), { injectIntoHead: head });
 
-        await forceHoveredState(page, '.hovered p-link-pure >>> a');
-        await forceHoveredState(page, '.hovered p-link-pure >>> span');
-        await forceFocusedState(page, '.focused p-link-pure >>> a');
-        await forceFocusedState(page, '.focused p-link-pure >>> span');
-        await forceFocusedHoveredState(page, '.focused-hovered p-link-pure >>> a');
-        await forceFocusedHoveredState(page, '.focused-hovered p-link-pure >>> span');
+        await forceHoveredState(page, '.hovered p-link-pure[href] >>> a');
+        await forceHoveredState(page, '.hovered p-link-pure:not([href]) >>> span'); // with slotted <a>, the shadowed <span> is used for hover styling
+        await forceFocusedState(page, '.focused p-link-pure[href] >>> a');
+        await forceFocusedState(page, '.focused:not([href]) p-link-pure a');
+        await forceFocusedHoveredState(page, '.focused-hovered p-link-pure[href] >>> a');
+        await forceFocusedState(page, '.focused-hovered p-link-pure:not([href]) a');
+        await forceHoveredState(page, '.focused-hovered p-link-pure:not([href]) >>> span'); // with slotted <a>, the shadowed <span> is used for hover styling
       })
     ).toBeFalsy();
   });