+++ packages/components-js/tests/vrt/specs/link-social.vrt.ts
@@ -30,7 +30,7 @@ describe('Link Social', () => {
       await vrt.test('link-social-states', async () => {
         const page = await vrt.getPage();
 
-        const head = `<style type="text/css">p-link-social { margin-right: 16px; margin-top: 16px; }</style>`;
+        const head = `<style type="text/css">p-link-social { margin-right: 1rem; margin-top: 1rem; }</style>`;
 
         const getElementsMarkup: GetThemedMarkup = (theme) => `
           <p-link-social theme="${theme}" href="https://www.porsche.com/">Fallback</p-link-social>
@@ -41,10 +41,18 @@ describe('Link Social', () => {
           <p-link-social theme="${theme}" href="https://www.google.com/" icon="logo-google" hide-label="true">Google</p-link-social>
           <p-link-social theme="${theme}" href="https://www.instagram.com/" icon="logo-instagram">Instagram</p-link-social>
           <p-link-social theme="${theme}" href="https://www.instagram.com/" icon="logo-instagram" hide-label="true">Instagram</p-link-social>
+          <p-link-social theme="${theme}" href="https://www.kakaocorp.com/" icon="logo-kakaotalk">KakaoTalk</p-link-social>
+          <p-link-social theme="${theme}" href="https://www.kakaocorp.com/" icon="logo-kakaotalk" hide-label="true">KakaoTalk</p-link-social>
           <p-link-social theme="${theme}" href="https://www.linkedin.com/" icon="logo-linkedin">LinkedIn</p-link-social>
           <p-link-social theme="${theme}" href="https://www.linkedin.com/" icon="logo-linkedin" hide-label="true">LinkedIn</p-link-social>
+          <p-link-social theme="${theme}" href="https://www.naver.com/" icon="logo-naver">Naver</p-link-social>
+          <p-link-social theme="${theme}" href="https://www.naver.com/" icon="logo-naver" hide-label="true">Naver</p-link-social>
           <p-link-social theme="${theme}" href="https://www.pinterest.com/" icon="logo-pinterest">Pinterest</p-link-social>
           <p-link-social theme="${theme}" href="https://www.pinterest.com/" icon="logo-pinterest" hide-label="true">Pinterest</p-link-social>
+          <p-link-social theme="${theme}" href="https://www.reddit.com/" icon="logo-reddit">Reddit</p-link-social>
+          <p-link-social theme="${theme}" href="https://www.reddit.com/" icon="logo-reddit" hide-label="true">Reddit</p-link-social>
+          <p-link-social theme="${theme}" href="https://www.tiktok.com/" icon="logo-tiktok">TikTok</p-link-social>
+          <p-link-social theme="${theme}" href="https://www.tiktok.com/" icon="logo-tiktok" hide-label="true">TikTok</p-link-social>
           <p-link-social theme="${theme}" href="https://www.twitter.com/" icon="logo-twitter">Twitter</p-link-social>
           <p-link-social theme="${theme}" href="https://www.twitter.com/" icon="logo-twitter" hide-label="true">Twitter</p-link-social>
           <p-link-social theme="${theme}" href="https://www.wechat.com/" icon="logo-wechat">Wechat</p-link-social>