+++ packages/components/src/components/form/select-wrapper/select-wrapper-styles.ts
@@ -1,9 +1,0 @@
-import { buildSlottedStyles, getBaseSlottedStyles, getCss, insertSlottedStyles } from '../../../utils';
-
-export const getSlottedCss = (host: HTMLElement): string => {
-  return getCss(buildSlottedStyles(host, getBaseSlottedStyles()));
-};
-
-export const addSlottedCss = (host: HTMLElement): void => {
-  insertSlottedStyles(host, getSlottedCss(host));
-};