+++ packages/components/CHANGELOG.md
@@ -8,6 +8,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### [Unreleased]
 
+#### Added
+- `Link Social` and `Icon` now support `kakaotalk`, `naver`, `reddit` and `tiktok`
+
+#### Changed
+- Alignment of `linkedin` icon
+
+#### Changed
+- Improved accessibility of `Select Wrapper`
+- `Icon` loading behaviour to non-blocking, components using the `Icon` will no longer wait for it to load
+- Validation messages of `Fieldset Wrapper` have now an additional icon representing the validation state
+
+#### Fixed
+- Box model of `Link Pure`
+- Focus of `Link Pure` with slotted anchor and hidden label
+– Focus cycling of `Modal` without focusable children
+
 ### [2.4.0-beta.1] - 2021-08-26
 
 #### Added
@@ -22,7 +38,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 #### Changed
 - Improved `:focus-visible` and `:hover:focus-visible` colors for `Link Social` and `Link`
 - Improved slotted `<a>` coloring in dark theme for `Link Social` and `Link`
-- Validation messages of `Checkbox Wrapper`, `Radio Button Wrapper`, `Select Wrapper`, `Textarea Wrapper` and `Text Field Wrapper` have now an additional icon representing the validation state.
+- Validation messages of `Checkbox Wrapper`, `Radio Button Wrapper`, `Select Wrapper`, `Textarea Wrapper` and `Text Field Wrapper` have now an additional icon representing the validation state
 - `Modal` backdrop behavior to close modal on mouse-down
 
 #### Fixed